+++ xrpl/queries/nft/v1/nftoken_buy_offers.go
@@ -19,6 +19,8 @@ type NFTokenBuyOffersRequest struct {
 	NFTokenID   types.NFTokenID        `json:"nft_id"`
 	LedgerHash  common.LedgerHash      `json:"ledger_hash,omitempty"`
 	LedgerIndex common.LedgerSpecifier `json:"ledger_index,omitempty"`
+	Limit       int                    `json:"limit,omitempty"`
+	Marker      any                    `json:"marker,omitempty"`
 }
 
 // Method returns the JSON-RPC method name for the NFTokenBuyOffersRequest.
@@ -45,4 +47,6 @@ func (*NFTokenBuyOffersRequest) Validate() error {
 type NFTokenBuyOffersResponse struct {
 	NFTokenID types.NFTokenID         `json:"nft_id"`
 	Offers    []nfttypes.NFTokenOffer `json:"offers"`
+	Limit     int                     `json:"limit,omitempty"`
+	Marker    any                     `json:"marker,omitempty"`
 }